Add title, image, authors and publisher table to books
I realized that it was a bad move to be heavily dependent on the
Google’s API because if traffic or the number of request is high then
the website won’t function like expected. Instead of fetching that
information from Google Servers 1 million time, I will just fetch it
one and save that data.