Create js module for page to sell your book
I want this page to refresh after a shipping label is created so that I
can redirect-off users to another page if the shipping label creation
was successful. However since I am opening the label in a new tab the
best way to do this is to refresh after a set timeout.

I chose 4sec because that should be enough time for the Shippo API to
give a response.