Delete tracking validator in AdminController@markAsPaid and change function order
I realized that updating the order, and then generating the shipping
label could be error prone because what if the label isn’t generated
and it throws an error? Now I have an Order instance that’s expected to
have a shipping entry but does not.

So I just moved the updating to a place that makes more sense.